Fix: use 4 bytes for bet_number

+++ services/client/src/client/bet.go
@@ -8,5 +8,5 @@ type Bet struct {
     Year        uint16
     Month       uint8
     Day         uint8
-    BetNumber   uint16
+    BetNumber   uint32
 }
